refactor: use trait instead of struct on lore_api_client from MailingListSelection
This commit refactors the way we store lore_api_client inside MailingListSelection struct.
The objective of this change is to make MailingListSelection functions testable. By using
BlockingLoreAPIClient directly in the struct it wasn't possible to instantiate a new
MailingListSelection for tests.
There is also a minor refactor within process_possible_mailing_lists to replace
a for loop with a more idiomatic iter + filter.
